Fix quota tests for static credentials
when running tempest tests with static set of credentials Tempest does
not create new tenants on the fly, and those pre-defined tenants may alreay
have configured quotas that are different from default project ones.

For quota tests to reliaby pass we need to store the original quotas in the
beginning of the test and update to them in the cleanUp phase instead
of deleting them (which resets them to the default ones).

Change-Id: Ia1de916a97fee81148976a5b9cb134e012af8e0c
Closes-Bug: #1800470