Improve the zone import API tests
This patch improves the zone import API tests such that they will
wait for a completed status before advancing. This not only insures
the test didn't get stuck in "PENDING", but also assures the import
record has been updated with the zone ID allowing the zone to be
cleaned up at test completion.

Change-Id: Ied4d4d37b3e5428f69b6ba2b0cfecb99d7125f2c