Unroll DDT API tests used for refstack/interop
This patch unrolls the test_create_all_recordset_types API tests from
using DDT. DDT is not OpenStack aware, so the idempotent_id's end up
being duplicated and the test names can be inconsistent due to value
padding.
This test has a limited number of combinations, so it is easy to unroll
these into stub tests that call a common, private, test method.

Closes-Bug: 1943115
Change-Id: I6d7c9358a336eedb313157ebad666d37d18f03d3