Fix blacklist scenario tests for scoped tokens
The blacklist scenario tests were added after the scoped token support
was added to designate-tempest-plugin. Since the scoped token jobs have
not yet merged in designate, we didn't catch that these tests were not
ready for scoped tokens.
This patch updates the test suite to use the system scoped admin token
when the tests are being run with scoped token support.

Change-Id: Ibfa282f6e54856cc9e9d4de1289de2c75e9db676