Fix API version test to not assume dns endpoint
The test_list_enabled_api_versions test was assuming the location of
the dns endpoint, but in some deployment that assumption is not true.
For example, if designate is using a specific port for the endpoint.

This patch corrects the test to use the keystone catalog to find the
proper designate API endpoint.

Change-Id: Ica672983214774d3ba38391d5fca6d42691f8e2a