Fix PTR API tests for scoped tokens
The PTR API tests were not using scoped tokens even though a test
job had scoped tokens enabled. This was not caught because the scoped
token patch for designate has not yet merged.
This patch updates the PTR API tests to use a scoped token for
admin if scoped tokens are enabled in the test run.

Change-Id: I259e7b94aae2e7794dd2514bb50cdc811c518845