Removal of tox envlist py3X
The removal of the envlist list value for unit test lessens the
confusion that the tempest test are tests in themselves and
hence we do not test the tests.

With the presents of the tox envlist, tox tries to run the tempest
tests as unit test, which results in failure and potential confusion.
The scope of tempest tests (Integration test) and unit test are
understandably different.

Change-Id: Ie016eca756c441cf557100c5227d640475885e4a