Fix test_list_enabled_api_versions test case
The test_list_enabled_api_versions was not properly checking the enabled API versions and would always fail on new API version additions.

This patch corrects the list comprehension and updates the test to look for at least one of the base API versions in the returned results. This way the test will not artificially fail as we add new API versions.

Change-Id: Ibcee798b15aaccf8fddd28f80df7112cda3a9ed5