Update the README file to use tox -e all
This patch updates the README file to have instructions for running the tempest tests using the "all" environment instead of "all-plugins". The "all-plugins" environment has been deprecated for years and uses site-packages, which may cause unintended test scenarios.
It also updates the example "run a single test" to be a non-admin API test.

Change-Id: I8645fb7577d52cd35e3d61d44ca9028ea3afe00f