Update hacking
This patch updates the version of hacking to be consistent with
designate.
It is required for flake8 to understand f-strings correctly.
Along with this version update, the patch fixes issues in the code found
by the new versions.

This patch also disables two flake8 plugins from tempest T110 and T111
as they are not compatible with the newer version of flake8.
They can be re-enabled once tempest updates to a newer hacking version
and fixes these issues.

Change-Id: Ia8153724e28cd0ce032d77ff08b4c35f75a1dfc8