Fix test_list_tsigkey_filter_by_algorithm scope parameter
The test_list_tsigkey_filter_by_algorithm test was creating a TSIG key
for a pool resource but omitting the scope parameter, which defaults to
'ZONE'. This caused the API to correctly reject the request with:

"Resource ID <pool-id> is not a valid zone"

This is a complementary patch to ensure the test correctly specifies
scope='POOL' when creating TSIG keys for pool resources (see [0]).

The API validation is working correctly by rejecting mismatched
scope/resource_id combinations. This patch fixes the test to properly
exercise the filter-by-algorithm functionality with a valid TSIG key.

[0]: https://review.opendev.org/c/openstack/designate/+/969084

Closes-Bug: #1934120
Change-Id: Ic186ae78acf22bb610e8e899f821868bc0306edf
Signed-off-by: Omer <oschwart@redhat.com>